A really large commit.
I forgot to add Bonus Cave 2 from the Godot version here, that will be in the next commit that will also add the Snake enemy and a new coin spritesheet. Welcome To PepperTux is not tested, so I don't recommend playing it. Kirby / The Pumpkin Enemy is also not tested, so I don't recommend using it in levels. Smartball is also not tested, so I don't recommend using it in levels.